add validation to user service
