+++ kitsune/Cargo.toml
@@ -48,7 +48,7 @@ garde = { version = "0.14.1", default-features = false, features = [
     "derive",
     "email",
     "email-idna",
-    "serde",
+    "regex",
 ] }
 globset = { version = "0.4.13", features = ["simd-accel"] }
 headers = "0.3.8"
@@ -122,7 +122,6 @@ zxcvbn = { version = "2.2.2", default-features = false }
 
 # "graphql" feature
 async-graphql = { version = "6.0.4", default-features = false, features = [
-    "email-validator",
     "playground",
     "tempfile",
     "time",
